properly modularise the tasks directory and add a basic readme
