+++ server/tasks/archiveInactiveThreads.js
@@ -1,0 +1,4 @@
+module.exports = async (logger) => {
+  console.log('This is the archive inactive message threads runner module');
+  logger.info('This is the archive inactive message threads runner module');
+};
