+++ server/tasks/exampleTaskModule.js
@@ -1,0 +1,4 @@
+module.exports = async (logger) => {
+  console.log('This is an example task runner module');
+  logger.info('This is an example task runner module');
+};
